+++ CHANGELOG.md
@@ -34,6 +34,8 @@ Note that since we don't clearly distinguish between a public and private interf
 - Renames PDB-Dev to PDB-IHM and adjusts data source
 - Support for custom Sequence Viewer mode options
 - Fix vertex based themes for spheres shader
+- Add volume dot representation
+- Add volume-value size theme
 - Sequence panel: Mark focused loci (bold+underline)
 - Change modifier key behavior in Normal Mode (default = select only, Ctrl/Cmd = add to selection, Shift = extend last selected range)